+++ app/src/main/java/com/maureen/wandevelop/feature/profile/bookmark/ui/DelBookmarkBottomSheet.kt
@@ -24,9 +24,9 @@ import com.maureen.wandevelop.R
 @OptIn(ExperimentalMaterial3Api::class)
 @Composable
 internal fun DelBookmarkBottomSheet(
-    bookmarkId: Long,
+    bookmarkId: Int,
     onCancel: () -> Unit,
-    onConfirm: (Long) -> Unit,
+    onConfirm: (Int) -> Unit,
     modifier: Modifier = Modifier
 ) {
     ModalBottomSheet(
